+++ lib/ahoy_email/engine.rb
@@ -4,12 +4,8 @@ module AhoyEmail
   class Engine < ::Rails::Engine
     initializer "ahoy_email" do |app|
       AhoyEmail.secret_token ||= begin
-        # Fix for issue with Mailkick and SECRET_KEY_BASE_DUMMY with Rails 7.1
-        # https://github.com/ankane/mailkick/pull/74
-        if Rails::VERSION::STRING.to_f >= 7.1 && ENV["SECRET_KEY_BASE_DUMMY"]
-          # TODO use for token in 3.0
-          app.key_generator.generate_key("ahoy_email")
-        end
+        tokens = []
+        tokens << app.key_generator.generate_key("ahoy_email")
 
         # TODO remove in 3.0
         creds =
@@ -23,7 +19,9 @@ module AhoyEmail
 
         token = creds.respond_to?(:secret_key_base) ? creds.secret_key_base : creds.secret_token
         token ||= app.secret_key_base # should come first, but need to maintain backward compatibility
-        token
+        tokens << token
+
+        tokens
       end
     end
   end