+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0 (unreleased)
 
 - Added support for key rotation
+- Improved secret token generation
 
 ## 2.2.0 (2023-07-02)